requirements.txt
bs4==0.0.1
python-dateutil==2.8.1
google-api-python-client==1.10.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.5
lxml==4.9.3
tendo==0.3.0
mojimoji==0.0.12